solved the db problem + extracted and displays artifact data + set welcome animation

+++ src/main/java/fr/uge/clone/Artefact.java
@@ -2,6 +2,6 @@ package fr.uge.clone;
 
 import java.sql.Date;
 
-public record Artefact(int id, String name, Date dateAdd, int analyzing, String url) {
+public record Artefact(long id, String name, Date dateAdd, int analyzing, String url) {
 
 }
